ZOOKEEPER-1675: Make sync a quorum operation
Previously, `sync` + `read` could not guarantee up-to-date data as
`sync` will not touch quorum in case of no outstanding proposals.

Though, `create`/`setData` could be used as an rescue, but it is
apparently ugly and error-prone. `sync` fits the semantics naturally.

This pr reverts ZOOKEEPER-2137 which using `setData` to circumvent no
quorum `sync`.

Since `sync` is a public API, so feature gate `quorumSync` is encouraged
to be off in rolling upgrade and turn on after rolling upgrade.

Refs: ZOOKEEPER-1675, ZOOKEEPER-2136, ZOOKEEPER-3600